requirements.txt
torch==2.5.1+cu118
torchvision==0.20.1+cu118
pyyaml==6.0.1
numpy==1.26.4
scipy==1.14.1
matplotlib==3.9.0